+++ src/main/java/net/mohron/skyclaims/permissions/Permissions.java
@@ -21,6 +21,7 @@ package net.mohron.skyclaims.permissions;
 public class Permissions {
 	// User Permissions
 	public static final String COMMAND_CREATE = "skyclaims.command.create";
+	public static final String COMMAND_EXPAND = "skyclaims.command.expand";
 	public static final String COMMAND_INFO = "skyclaims.command.info";
 	public static final String COMMAND_INVITE = "skyclaims.command.invite";
 	public static final String COMMAND_LIST = "skyclaims.command.list";