+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Change Log
 
 ## Beta 17
+**REQUIRES: SF 2096+ & GP 255+ (latest versions are highly recommended)**
 - Fixed island width calculation being off by 2 `/is info`
+- Added `is expand [blocks]` to allow player to expand their islands using claim blocks _(subtracted from accrued blocks - will regen)_
+- Added `skyclaims.max-size` option to limit how large players can expand their islands
 - `skyclaims.initial-size` now supports being set to 256
 - Added a confirmation message to the admin delete shortcut in `/is info`
-- Remove hard island limit from `isa transfer` (multiple islands per player is still unsupported officially)
+- Remove hard island limit from `isa transfer` _(multiple islands per player is still unsupported officially)_
 
 ## Beta 16
 **REQUIRES: SF 2096+ & GP 255+ (latest versions are highly recommended)**